Split feature toggle view in smaller components
